+++ tests/dummy/snippets/blacklist.js
@@ -1,0 +1,6 @@
+ENV['ember-paper'] = {
+    blacklist: [
+        'paper-card',
+        'paper-item'
+    ]
+};