+++ tests/dummy/snippets/whitelist.js
@@ -1,0 +1,7 @@
+ENV['ember-paper'] = {
+    whitelist: [
+        'paper-card',
+        'paper-tooltip',
+        'paper-button'
+    ]
+};