+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - [#1024](https://github.com/miguelcobain/ember-paper/pull/1024) Associate input field with corresponding info messages for better a11y
 - [#1026](https://github.com/miguelcobain/ember-paper/pull/1026) support `aria-hidden` argument on paper-icon
 - [#588](https://github.com/miguelcobain/ember-paper/pull/588) Allow paper-autocomplete to be blockless
+- [#1021](https://github.com/miguelcobain/ember-paper/pull/1021) you can now whitelist/blacklist components
 
 ### 1.0.0-beta.18
 - [68ca72e](https://github.com/miguelcobain/ember-paper/commit/d931d1918e9ddd208cecf045bb6378b39148d7d1) not using ember-cli-sass-variables-export addon anymore since it doesn't output the generated files when using `ember build` (just `ember s`). We're manually including the generated files until we find a better solution. The palettes aren't likely to change anyway.